+++ crates/amaru-ledger/src/startup.rs
@@ -12,21 +12,47 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-use amaru_kernel::{Epoch, EraHistory, GovernanceAction, ProtocolParameters};
-use amaru_observability::info;
+use std::ops::Deref;
 
-use crate::store::{self, ReadStore, StoreError};
+use amaru_kernel::{Credential, Epoch, EraHistory, GovernanceAction, ProtocolParameters, hash};
+use amaru_observability::{info, info_span};
+use tracing::field;
 
-pub type StartupHook<S> = fn(&Database<'_, S>) -> Result<(), StoreError>;
+use crate::store::{ReadStore, StoreError};
 
-pub struct Database<'a, S: ReadStore> {
+// ------------------------------------------------------------------------------------- StartupHook
+
+pub type StartupHook<S> = fn(&StartupContext<'_, S>) -> Result<(), StoreError>;
+
+pub fn no_startup_hook<S: ReadStore>(_: &StartupContext<'_, S>) -> Result<(), StoreError> {
+    Ok(())
+}
+
+pub fn with_startup_hook<'a, S: ReadStore>(ctx: &StartupContext<'a, S>) -> Result<(), StoreError> {
+    ctx.emit_protocol_parameters();
+    ctx.emit_current_pots()?;
+    ctx.emit_active_proposals()?;
+    ctx.emit_constitutional_committee()?;
+    Ok(())
+}
+
+// ---------------------------------------------------------------------------------- StartupContext
+
+pub struct StartupContext<'a, S: ReadStore> {
     stable: &'a S,
     epoch: Epoch,
     protocol_parameters: &'a ProtocolParameters,
     era_history: &'a EraHistory,
 }
 
-impl<'a, S: ReadStore> Database<'a, S> {
+impl<'a, S: ReadStore> Deref for StartupContext<'a, S> {
+    type Target = S;
+    fn deref(&self) -> &Self::Target {
+        self.stable
+    }
+}
+
+impl<'a, S: ReadStore> StartupContext<'a, S> {
     pub(crate) fn new(
         stable: &'a S,
         epoch: Epoch,
@@ -48,97 +74,184 @@ impl<'a, S: ReadStore> Database<'a, S> {
         self.era_history
     }
 
-    pub fn pots(&self) -> Result<store::columns::pots::Row, StoreError> {
-        self.stable.pots()
+    fn emit_protocol_parameters(&self) {
+        let ProtocolParameters {
+            protocol_version,
+            max_block_body_size,
+            max_transaction_size,
+            max_block_header_size,
+            max_tx_ex_units,
+            max_block_ex_units,
+            max_value_size,
+            max_collateral_inputs,
+            min_fee_a,
+            min_fee_b,
+            stake_credential_deposit,
+            stake_pool_deposit,
+            monetary_expansion_rate,
+            treasury_expansion_rate,
+            min_pool_cost,
+            lovelace_per_utxo_byte,
+            prices,
+            min_fee_ref_script_lovelace_per_byte,
+            max_ref_script_size_per_tx,
+            max_ref_script_size_per_block,
+            ref_script_cost_stride,
+            ref_script_cost_multiplier,
+            stake_pool_max_retirement_epoch,
+            optimal_stake_pools_count,
+            pledge_influence,
+            collateral_percentage,
+            cost_models: _,
+            pool_voting_thresholds,
+            drep_voting_thresholds,
+            min_committee_size,
+            max_committee_term_length,
+            gov_action_lifetime,
+            gov_action_deposit,
+            drep_deposit,
+            drep_expiry,
+        } = &self.protocol_parameters;
+
+        info!(
+            ledger::protocol_parameters::DUMP,
+            protocol_version,
+            max_block_body_size,
+            max_transaction_size,
+            max_block_header_size,
+            max_tx_ex_units,
+            max_block_ex_units,
+            max_value_size,
+            max_collateral_inputs,
+            min_fee_a,
+            min_fee_b,
+            stake_credential_deposit,
+            stake_pool_deposit,
+            monetary_expansion_rate,
+            treasury_expansion_rate,
+            min_pool_cost,
+            lovelace_per_utxo_byte,
+            prices,
+            min_fee_ref_script_lovelace_per_byte,
+            max_ref_script_size_per_tx,
+            max_ref_script_size_per_block,
+            ref_script_cost_stride,
+            ref_script_cost_multiplier,
+            stake_pool_max_retirement_epoch,
+            optimal_stake_pools_count,
+            pledge_influence,
+            collateral_percentage,
+            pool_voting_thresholds,
+            drep_voting_thresholds,
+            min_committee_size,
+            max_committee_term_length,
+            gov_action_lifetime,
+            gov_action_deposit,
+            drep_deposit,
+            drep_expiry,
+        );
     }
 
-    pub fn iter_proposals(
-        &self,
-    ) -> Result<impl Iterator<Item = (store::columns::proposals::Key, store::columns::proposals::Row)> + '_, StoreError>
-    {
-        self.stable.iter_proposals()
+    fn emit_current_pots(&self) -> Result<(), StoreError> {
+        let pots = self.pots()?;
+
+        info!(
+            ledger::pots::DUMP,
+            treasury = pots.treasury,
+            reserves = pots.reserves,
+            fees = pots.fees,
+            donations = pots.donations,
+        );
+
+        Ok(())
     }
-}
 
-pub fn with_startup_hook<S: ReadStore>(database: &Database<'_, S>) -> Result<(), StoreError> {
-    emit_protocol_parameters(database);
-    emit_current_pots(database)?;
-    emit_active_proposals(database)?;
-    Ok(())
-}
+    fn emit_active_proposals(&self) -> Result<(), StoreError> {
+        for (id, row) in self.iter_proposals()? {
+            let proposal_kind = proposal_kind(&row.proposal.gov_action);
+            let detail = proposal_detail(&row.proposal.gov_action);
+            let proposed_in = self
+                .era_history()
+                .slot_to_epoch_unchecked_horizon(row.proposed_in.transaction.slot)
+                .map_err(|error| StoreError::Internal(Box::new(error)))?;
 
-pub fn no_startup_hook<S: ReadStore>(_: &Database<'_, S>) -> Result<(), StoreError> {
-    Ok(())
-}
+            if let Some(detail) = detail {
+                info!(
+                    ledger::proposal::ACTIVE,
+                    id = id.to_string(),
+                    proposal_kind,
+                    proposed_in,
+                    valid_until = row.valid_until,
+                    detail = @detail,
+                );
+            } else {
+                info!(
+                    ledger::proposal::ACTIVE,
+                    id = id.to_string(),
+                    proposal_kind,
+                    proposed_in,
+                    valid_until = row.valid_until,
+                );
+            }
+        }
 
-fn emit_protocol_parameters<S: ReadStore>(database: &Database<'_, S>) {
-    let protocol_parameters = database.protocol_parameters();
+        Ok(())
+    }
 
-    info!(
-        ledger::protocol_parameters::LOAD,
-        protocol_version = protocol_parameters.protocol_version.to_string(),
-        max_block_body_size = protocol_parameters.max_block_body_size.to_string(),
-        max_transaction_size = protocol_parameters.max_transaction_size.to_string(),
-        max_tx_ex_units = protocol_parameters.max_tx_ex_units.to_string(),
-        max_block_ex_units = protocol_parameters.max_block_ex_units.to_string(),
-        min_fee_a = protocol_parameters.min_fee_a.to_string(),
-        min_fee_b = protocol_parameters.min_fee_b.to_string(),
-        stake_credential_deposit = protocol_parameters.stake_credential_deposit.to_string(),
-        stake_pool_deposit = protocol_parameters.stake_pool_deposit.to_string(),
-        lovelace_per_utxo_byte = protocol_parameters.lovelace_per_utxo_byte.to_string(),
-        collateral_percentage = protocol_parameters.collateral_percentage.to_string(),
-        gov_action_lifetime = protocol_parameters.gov_action_lifetime.to_string(),
-        gov_action_deposit = protocol_parameters.gov_action_deposit.to_string(),
-        drep_deposit = protocol_parameters.drep_deposit.to_string(),
-        drep_expiry = protocol_parameters.drep_expiry.to_string(),
-    );
-}
+    #[expect(clippy::panic)]
+    fn emit_constitutional_committee(&self) -> Result<(), StoreError> {
+        info_span!(ledger::constitutional_committee::DUMP, status = self.constitutional_committee()?);
 
-fn emit_current_pots<S: ReadStore>(database: &Database<'_, S>) -> Result<(), StoreError> {
-    let pots = database.pots()?;
+        fn is_invalid_cc_member(cold_credential: &Credential) -> bool {
+            false
+                || cold_credential
+                    == &Credential::ScriptHash(hash!("349e55f83e9af24813e6cb368df6a80d38951b2a334dfcdf26815558"))
+                || cold_credential
+                    == &Credential::ScriptHash(hash!("9cc3f387623f45dae6a68b7096b0c2e403d8601a82dc40221ead41e2"))
+                || cold_credential
+                    == &Credential::KeyHash(hash!("dc0d6ef49590eb6880a50a00adde17596e6d76f7159572fa1ff85f2a"))
+        }
 
-    info!(
-        ledger::pots::LOAD,
-        treasury = pots.treasury,
-        reserves = pots.reserves,
-        fees = pots.fees,
-        donations = pots.donations,
-    );
+        for (cold_credential, member) in self.iter_cc_members()? {
+            if self.epoch == Epoch::from(654) && is_invalid_cc_member(&cold_credential) {
+                panic!(
+                    r#"
+    Corrupted ledger state: bootstrap or manual rollback is required.
 
-    Ok(())
-}
+    Amaru versions prior to v10.11.20260912 contains an off-by-one error which
+    prevented the correct ratification of CC members at the boundary of 653→654.
 
-fn emit_active_proposals<S: ReadStore>(database: &Database<'_, S>) -> Result<(), StoreError> {
-    for (id, row) in database.iter_proposals()? {
-        let proposal_kind = proposal_kind(&row.proposal.gov_action);
-        let detail = proposal_detail(&row.proposal.gov_action);
-        let proposed_in = database
-            .era_history()
-            .slot_to_epoch_unchecked_horizon(row.proposed_in.transaction.slot)
-            .map_err(|error| StoreError::Internal(Box::new(error)))?;
+    Fixing this requires to either:
 
-        if let Some(detail) = detail {
+    - rollback manually if you still have the appropriate ledger states:
+
+      ```
+      amaru node rollback --epoch 654
+      ```
+
+    - re-bootstrap your node:
+
+      ```
+      amaru node rm --wipe-all-dbs
+      amaru node bootstrap
+      ```
+
+    (sorry for the inconvenience :s)"#
+                );
+            }
+
             info!(
-                ledger::proposal::ACTIVE,
-                id = id.to_string(),
-                proposal_kind,
-                proposed_in,
-                valid_until = row.valid_until,
-                detail,
-            );
-        } else {
-            info!(
-                ledger::proposal::ACTIVE,
-                id = id.to_string(),
-                proposal_kind,
-                proposed_in,
-                valid_until = row.valid_until,
+                ledger::constitutional_committee_member::DUMP,
+                cold_credential = cold_credential,
+                status = @member.status.as_ref().map(field::display),
+                valid_until = @member.valid_until.as_ref().map(|epoch| epoch.as_u64()),
             );
         }
+        Ok(())
     }
-
-    Ok(())
 }
+
+// ----------------------------------------------------------------------------------------- Helpers
 
 fn proposal_kind(proposal: &GovernanceAction) -> &'static str {
     match proposal {